feat(reactor-browser): split analytics into store and react hooks
Move createAnalyticsStore/getAnalyticsStore to store.ts with no React
or tanstack dependency. Keep React hooks and Provider in context.tsx.
Export only store.ts from the root index to eliminate @tanstack/react-query
from the root bundle. Analytics hooks remain available via the /analytics
subpath export.
